requirements.txt
mcp[cli]==1.10.1
google-generativeai>=0.7.0
python-dotenv
ipykernel
httpx
pydantic
fastapi
uvicorn
requests